opt/cs50/submit50/bin/requirements.txt
getch
github3.py
termcolor
